Change SzabadnapokSzama type from ushort to byte
Updated the SzabadnapokSzama property in EmployeeViewmodel to use byte instead of ushort, reducing its value range to 0-255. This change may help optimize memory usage if higher values are not needed.

+++ WMSProject/WMSProject/EmployeeViewmodel.cs
@@ -4,7 +4,7 @@
     {
         public string Nev { get; } = "Nagy Anna";
         public string Adoszam { get; } = "1234567890-123";
-        public ushort SzabadnapokSzama { get; } = 15;
+        public byte SzabadnapokSzama { get; } = 15;
         public EmployeeViewmodel()
         {
         }
